feat: WebSocket 프레임 처리 및 화재 감지 로직 통합
- WebSocket을 통해 수신한 프레임 데이터의 유효성을 검증하는 로직을 video_processing_service로 통합
- Base64 인코딩된 이미지를 디코딩하고 화재 감지를 수행하는 analyze_frame_for_fire 함수 추가
- 화재 감지 시 알림 전송을 위한 send_fire_alert 함수 호출 로직 구현
- 프레임 처리 루프에서 JSON 데이터 파싱 및 오류 처리 개선

이제 WebSocket을 통해 실시간으로 화재 감지 결과를 처리하고 알림을 전송할 수 있습니다.
